feat: home optionでdotfileを置くディレクトリを指定できるように変更

+++ setup.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
 DOT_FILES=()
-CURRNT_DIR=$(cd $(dirname $0) && pwd)
+HOME=$HOME
+CURRENT_DIR=$(cd $(dirname $0) && pwd)
 
-while getopts ptvz-: o; do
+while getopts tvz-: o; do
   case "$o" in
     -)
       case "${OPTARG}" in
@@ -11,6 +12,13 @@ while getopts ptvz-: o; do
           DOT_FILES=('.tmux.conf .vimrc .zshrc')
           vi=true
           zsh=true;;
+        home)
+          HOME=${BASH_ARGV}
+          if [ ! -e $HOME ]; then
+          echo "$HOME is not exist"; exit 1
+          fi
+          #cd $HOME > /dev/null 2>&1 || echo "$HOME is not exist"; exit 1
+          ;;
         tmux)
           DOT_FILES=('.tmux.conf' "${DOT_FILES[@]}")
           ;;
@@ -35,29 +43,31 @@ done
 
 for dotfile in "${DOT_FILES[@]}"
 do
-  mv $HOME/$dotfile $HOME/$dotfile.old
+  if [ -e $HOME/$dotfile ]; then
+    mv $HOME/$dotfile $HOME/$dotfile.old
+  fi
   cp $dotfile $HOME
 done
 
 if [ $zsh ]; then
-  if [ ! -e $CURRNT_DIR/zsh-syntax-highlighting ]; then
-    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $CURRNT_DIR/zsh-syntax-highlighting
+  if [ ! -e $CURRENT_DIR/zsh-syntax-highlighting ]; then
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git $CURRENT_DIR/zsh-syntax-highlighting
   fi
 
-  if [ ! -e $CURRNT_DIR/zsh-interactive-cd ]; then
-    git clone https://github.com/changyuheng/zsh-interactive-cd.git $CURRNT_DIR/zsh-interactive-cd
+  if [ ! -e $CURRENT_DIR/zsh-interactive-cd ]; then
+    git clone https://github.com/changyuheng/zsh-interactive-cd.git $CURRENT_DIR/zsh-interactive-cd
   fi
 
-  echo "source $CURRNT_DIR/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" >> $HOME/.zshrc
-  echo "source $CURRNT_DIR/zsh-interactive-cd/zsh-interactive-cd.plugin.zsh" >> $HOME/.zshrc
-  echo "source ~/.fzf.zsh" >> $HOME/.zshrc
+  echo "source $CURRENT_DIR/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" >> $HOME/.zshrc
+  echo "source $CURRENT_DIR/zsh-interactive-cd/zsh-interactive-cd.plugin.zsh" >> $HOME/.zshrc
+  echo "source $HOME/.fzf.zsh" >> $HOME/.zshrc
   # echo "source ~/.fzf/shell/key-bindings.zsh" >> $HOME/.zshrc
-  source ~/.zshrc
+  source $HOME/.zshrc
 fi
 
 if [ $vi ]; then
   if [ ! -e $HOME/.vim/toml ]; then
-    mkdir -p ~/.vim/toml
+    mkdir -p $HOME/.vim/toml
   fi
-  cp -f ./dein.toml $HOME/.vim/toml/
+  cp -f $CURRENT_DIR/dein.toml $HOME/.vim/toml/
 fi
